HW. Created data.js and imported to main.js
